pyproject.toml
[project]
name = "mundo-auto"
version = "0.1.0"
description = "E-commerce system for auto parts"
authors = [
    {name = "Pedro Domingos"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "pydantic[email] (>=2.12.2,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "google-auth (>=2.41.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
black = "^25.9.0"
isort = "^7.0.0"
mypy = "^1.18.2"

